refactor: file structure - processors in separate module
