+++ crates/worker/src/commands/errors.rs
@@ -1,5 +1,3 @@
-use crate::clients::errors::{DBError, S3Error};
-
 #[derive(thiserror::Error, Debug)]
 pub enum CompilationError {
     #[error("Failed to compile: {0}")]
