+++ crates/worker/src/errors.rs
@@ -1,7 +1,6 @@
 use types::item::ItemError;
 
 use crate::clients::errors::{DBError, S3Error};
-use crate::commands::errors::CompilationError;
 
 #[derive(thiserror::Error, Debug)]
 pub enum PurgeError {
@@ -11,13 +10,4 @@ pub enum PurgeError {
     S3Error(#[from] S3Error),
     #[error("ItemError: {0}")]
     ItemError(#[from] ItemError),
-}
-#[derive(thiserror::Error, Debug)]
-pub enum CompileProcessorError {
-    #[error("Unsupported version: {0}")]
-    VersionNotSupportedError(String),
-    #[error("CompilationError: {0}")]
-    CompilationError(#[from] CompilationError),
-    #[error("UnknownError: {0}")]
-    UnknownError(#[from] anyhow::Error),
 }
