+++ crates/worker/src/processor/compile_processor.rs
@@ -1,11 +1,3 @@
-use crate::clients::s3_clients::wrapper::S3ClientWrapper;
-use crate::clients::sqs_clients::wrapper::SqsClientWrapper;
-use crate::commands::compile::{do_compile, CompilationOutput};
-use crate::errors::CompileProcessorError;
-use crate::input_preparator::InputPreparator;
-use crate::purgatory::Purgatory;
-use crate::utils::cleaner::AutoCleanUp;
-use crate::utils::lib::{s3_compilation_files_dir, ZKSOLC_VERSIONS};
 use anyhow::Context;
 use aws_sdk_s3::presigning::PresigningConfig;
 use std::time::Duration;
@@ -13,6 +5,15 @@ use tracing::error;
 use types::item::TaskResult;
 use types::{CompilationRequest, ARTIFACTS_FOLDER};
 use uuid::Uuid;
+
+use crate::clients::s3_clients::wrapper::S3ClientWrapper;
+use crate::clients::sqs_clients::wrapper::SqsClientWrapper;
+use crate::commands::compile::{do_compile, CompilationOutput};
+use crate::processor::errors::CompileProcessorError;
+use crate::processor::input_preparator::InputPreparator;
+use crate::purgatory::Purgatory;
+use crate::utils::cleaner::AutoCleanUp;
+use crate::utils::lib::{s3_compilation_files_dir, ZKSOLC_VERSIONS};
 
 pub struct CompileProcessor {
     sqs_client: SqsClientWrapper,
