+++ crates/worker/src/processor/input_preparator.rs
